# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ChangePasswordRequest": {
        "additionalProperties": false,
        "description": "The current password and the replacement.\n\nThe account is the one the access token belongs to; this body cannot name a user.\nA new password that breaks the password policy is answered with `422` saying which\nrule it broke.",
        "properties": {
          "current_password": {
            "maxLength": 1024,
            "minLength": 1,
            "title": "Current Password",
            "type": "string"
          },
          "new_password": {
            "maxLength": 1024,
            "minLength": 1,
            "title": "New Password",
            "type": "string"
          }
        },
        "required": [
          "current_password",
          "new_password"
        ],
        "title": "ChangePasswordRequest",
        "type": "object"
      },
...
      "CurrentUserResponse": {
        "description": "Fields are enumerated rather than serialised from the entity.\n\n`password_hash` lives on `User`, so a `from_attributes` dump would leak it. Also\nomitted: `status`, which is an internal detail the caller does not need (R2.6).",
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "must_change_password": {
            "title": "Must Change Password",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "preferred_language": {
            "title": "Preferred Language",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "tenant_id": {
            "format": "uuid",
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenant_id",
          "name",
          "email",
          "role",
          "preferred_language",
          "must_change_password"
        ],
        "title": "CurrentUserResponse",
        "type": "object"
      },
...
      "ErrorBody": {
        "description": "The `error` object of the PRD §23 envelope.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorCode": {
        "enum": [
          "INTERNAL_ERROR",
          "HTTP_ERROR",
          "VALIDATION_ERROR",
          "CONFLICT",
          "PAYLOAD_TOO_LARGE",
          "METHOD_NOT_ALLOWED",
          "INVALID_CREDENTIALS",
          "INVALID_TOKEN",
          "FORBIDDEN",
          "RATE_LIMITED",
          "PASSWORD_CHANGE_REQUIRED",
          "NOT_FOUND",
          "BAD_GATEWAY"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "ErrorEnvelope": {
        "description": "Mirror of `app.core.errors.error_envelope()` — the only error shape this API emits.\n\n`code` is typed as `ErrorCode` on purpose (design D11): it publishes the catalogue of\ncodes as an enum, so a consumer can switch over it exhaustively with its compiler's\nhelp. `app/core/error_codes.py` is the single source of truth that keeps that\ncatalogue from going stale.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "ForgotPasswordRequest": {
        "additionalProperties": false,
        "description": "The address to send a recovery link to.",
        "properties": {
          "email": {
            "maxLength": 255,
            "minLength": 3,
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "title": "ForgotPasswordRequest",
        "type": "object"
      },
      "ForgotPasswordResponse": {
        "description": "The same acknowledgement for every request.\n\nIt does not indicate whether the address belongs to an account.",
        "properties": {
          "detail": {
            "default": "If the address belongs to an account, a recovery link has been sent.",
            "title": "Detail",
            "type": "string"
          }
        },
        "title": "ForgotPasswordResponse",
        "type": "object"
      },
...
      "LoginRequest": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "maxLength": 255,
            "minLength": 3,
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 1024,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
...
      "RefreshRequest": {
        "additionalProperties": false,
        "properties": {
          "refresh_token": {
            "minLength": 1,
            "title": "Refresh Token",
            "type": "string"
          }
        },
        "required": [
          "refresh_token"
        ],
        "title": "RefreshRequest",
        "type": "object"
      },
...
      "ResetPasswordRequest": {
        "additionalProperties": false,
        "description": "The recovery token from the emailed link, and the new password.",
        "properties": {
          "new_password": {
            "maxLength": 1024,
            "minLength": 1,
            "title": "New Password",
            "type": "string"
          },
          "token": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token",
          "new_password"
        ],
        "title": "ResetPasswordRequest",
        "type": "object"
      },
...
      "TokenPairResponse": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "refresh_token": {
            "title": "Refresh Token",
            "type": "string"
          },
          "token_type": {
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "refresh_token",
          "token_type",
          "expires_in"
        ],
        "title": "TokenPairResponse",
        "type": "object"
      },
...
      "UserRole": {
        "enum": [
          "SUPER_ADMIN",
          "TENANT_OWNER",
          "PROPERTY_MANAGER",
          "CLEANER",
          "TECHNICIAN"
        ],
        "title": "UserRole",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "AutoHostAI backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/change-password": {
      "post": {
        "description": "Self-service, for any role that can authenticate. The subject is the holder of the access token — the body cannot name another user. On success EVERY refresh family of the account is revoked, including the one that made this call, so the caller must log in again: a change that left the old sessions alive would add a credential rather than rotate one.",
        "operationId": "change_password_api_v1_auth_change_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Missing, malformed or expired credentials."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authenticated, but the role lacks the required permission."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Change your own password",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/forgot-password": {
      "post": {
        "description": "Anonymous. Always answers the same 202 with the same body, whether or not the address belongs to an account — otherwise it would be a user-enumerator open to the internet. Shares the per-IP rate limit with login and refresh.",
        "operationId": "forgot_password_api_v1_auth_forgot_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForgotPasswordResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ask for a password recovery link",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "description": "Anonymous. Rate limited per client IP, and an account is temporarily locked after too many consecutive failures. Every failure answers the same 401, whatever the cause.",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPairResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Exchange email and password for a token pair",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "description": "Revokes the refresh family named by the token. Access tokens already issued keep working until they expire — at most their configured lifetime.",
        "operationId": "logout_api_v1_auth_logout_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Missing, malformed or expired credentials."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authenticated, but the role lacks the required permission."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "End the session this access token belongs to",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "description": "Returns the identity behind the access token: the user, their role and the tenant the token is scoped to. Reads nothing from the request beyond the token, so it never resolves another tenant's data.",
        "operationId": "me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Missing, malformed or expired credentials."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authenticated, but the role lacks the required permission."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "The authenticated user's own profile",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "description": "Anonymous: the refresh token itself is the credential. The presented token is invalidated. Presenting an already-used one revokes the whole session family.",
        "operationId": "refresh_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPairResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rotate a refresh token",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/reset-password": {
      "post": {
        "description": "Anonymous: the token from the emailed link is the credential, and it is single use. Every reason it can fail — unknown, already used, expired, revoked, or an account that is no longer active — answers the same 401. On success every session of the account is revoked and any other outstanding recovery link is invalidated; no session is returned, so log in afterwards. Shares the per-IP rate limit with login and refresh.",
        "operationId": "reset_password_api_v1_auth_reset_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set a new password using a recovery token",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}